+++ Dimitri/Game/Game/src/Models/Abstract/Interactable.cpp
@@ -1,6 +1,6 @@
 #include "Interactable.h"
 using namespace Game;
 
-Models::Interactable::Interactable(int x, int y, int z, int height, int width, Enums::DirectionEnum state, Graphics::Models::Center center) : Models::Object(x, y, z, height, width, state, center)
+Models::Interactable::Interactable(float x, float y, float z, float height, float width, Enums::DirectionEnum state, Graphics::Models::Center center) : Models::Object(x, y, z, height, width, state, center)
 {
 }
